added online and offline/ authorized and not authorized to favorite screen

+++ app/src/main/java/io/github/boodyahmedhamdy/mealano/favorite/presenter/FavoritePresenter.java
@@ -10,6 +10,7 @@ import io.github.boodyahmedhamdy.mealano.datalayer.datasources.local.db.entities
 import io.github.boodyahmedhamdy.mealano.datalayer.repos.MealsRepository;
 import io.github.boodyahmedhamdy.mealano.datalayer.repos.UsersRepository;
 import io.github.boodyahmedhamdy.mealano.favorite.contract.FavoriteView;
+import io.github.boodyahmedhamdy.mealano.utils.network.NetworkMonitor;
 import io.github.boodyahmedhamdy.mealano.utils.rx.OnBackgroundTransformer;
 import io.reactivex.rxjava3.disposables.Disposable;
 
@@ -20,32 +21,47 @@ public class FavoritePresenter {
     MealsRepository mealsRepository;
     UsersRepository usersRepository;
 
-    public FavoritePresenter(FavoriteView view, MealsRepository mealsRepository, UsersRepository usersRepository) {
+    NetworkMonitor networkMonitor;
+
+
+    public FavoritePresenter(FavoriteView view, MealsRepository mealsRepository, UsersRepository usersRepository, NetworkMonitor networkMonitor) {
         this.view = view;
         this.mealsRepository = mealsRepository;
         this.usersRepository = usersRepository;
+        this.networkMonitor = networkMonitor;
     }
 
     public void getFavoriteMeals() {
+        if(usersRepository.getCurrentUser() != null) {
+            view.setIsAuthorized(true);
 
-        Disposable dis =  mealsRepository.getFavoriteMeals(
-                usersRepository.getCurrentUser().getUid()
-        ).compose(new OnBackgroundTransformer<>())
-                .subscribe(mealEntities -> {
-                    view.setFavoriteMeals(mealEntities);
-                }, throwable -> {
-                    view.setErrorMessage(throwable.getLocalizedMessage());
-                });
+            Disposable dis =  mealsRepository.getFavoriteMeals(
+                    usersRepository.getCurrentUser().getUid()
+            ).compose(new OnBackgroundTransformer<>())
+                    .subscribe(mealEntities -> {
+                        view.setFavoriteMeals(mealEntities);
+                    }, throwable -> {
+                        view.setErrorMessage(throwable.getLocalizedMessage());
+                    });
+        } else {
+            view.setIsAuthorized(false);
+        }
 
     }
 
     public void deleteFavoriteMeal(MealEntity meal) {
-        Disposable dis =  mealsRepository.deleteFavoriteMeal(meal)
-                .compose(new OnBackgroundTransformer<>())
-                .subscribe(
-                        () -> Log.i(TAG, "deleteFavoriteMeal: deleted " + meal.mealDTO.getStrMeal() + " successfully"),
-                        throwable -> Log.e(TAG, "deleteFavoriteMeal: ", throwable)
-                );
+        if(networkMonitor.isConnected()) {
+            view.setIsOnline(true);
+            Disposable dis =  mealsRepository.deleteFavoriteMeal(meal)
+                    .compose(new OnBackgroundTransformer<>())
+                    .subscribe(
+                            () -> Log.i(TAG, "deleteFavoriteMeal: deleted " + meal.mealDTO.getStrMeal() + " successfully"),
+                            throwable -> Log.e(TAG, "deleteFavoriteMeal: ", throwable)
+                    );
+        } else {
+            view.setIsOnline(false);
+        }
+
 
     }
 }
